+++ src/main/java/pl/tw/ksiegowosc/config/MeritApiProperties.java
@@ -8,8 +8,6 @@ import org.springframework.validation.annotation.Validated;
 @ConfigurationProperties(prefix = "clients.merit")
 public record MeritApiProperties(
         @NotBlank String baseUrl,
-        @NotBlank String apiId,
-        @NotBlank String apiKey,
         @NotBlank String v2BaseUrl
 ) {
 }