+++ src/main/java/pl/tw/ksiegowosc/dto/MeritCredentials.java
@@ -1,0 +1,7 @@
+package pl.tw.ksiegowosc.dto;
+
+public record MeritCredentials(
+        String apiId,
+        String apiKey
+) {
+}